+++ components/ReusableSmallComponents/iconsOrSvgImages/PawPrintIcon.tsx
@@ -1,8 +1,10 @@
+/**
+ * Font Awesome paw icon for list bullets.
+ * Notes: docs/notes/components/reusable-small-components.md
+ */
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faPaw } from "@fortawesome/free-solid-svg-icons";
 
-const PawPrintIcon = () => {
+export default function PawPrintIcon() {
   return <FontAwesomeIcon icon={faPaw} className="yellow mr-2" />;
-};
-
-export default PawPrintIcon;
+}